requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
PyMySQL==1.1.1
requests==2.32.3
google-auth==2.33.0
python-dotenv==1.0.1
